# src/main/resources/swagger/coupon-api.yml
openapi: 3.0.1
info:
  title: Coupons API
  version: "1.0.0"

paths:
  /coupons:
    get:
      operationId: getAllCoupons
      description: Endpoint for fetching all coupons without pagination
      responses:
        '200':
          description: Created coupon response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CouponData'
    post:
      operationId: createCoupon
      description: Endpoint for creating new coupon
      requestBody:
        content:
          application/json:
            schema:
              title: CreateCouponBody
              type: object
              properties:
                code:
                  type: string
                  minLength: 1
                usageLimit:
                  type: integer
                  minimum: 1
                countryCode:
                  type: string
                  minLength: 1
                  maxLength: 2
      responses:
        '200':
          description: Created coupon response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicCouponData'
        '500':
          $ref: "#/components/responses/InternalServerError"
        '400':
          $ref: '#/components/responses/BadRequest'
  /coupons/{code}/register:
    post:
      operationId: registerUserCoupon
      description: Endpoint for registering user coupons
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              title: RegisterCouponBody
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Registered coupon response
        '500':
          $ref: "#/components/responses/InternalServerError"
        '400':
          $ref: '#/components/responses/BadRequest'
components:
  schemas:
    BasicCouponData:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        createDate:
          type: string
          format: date-time
        usageLimit:
          type: integer
        countryCode:
          type: string
    CouponData:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        createDate:
          type: string
          format: date-time
        usageLimit:
          type: integer
        currentUsage:
          type: integer
        countryCode:
          type: string
  responses:
    Ok:
      description: OK
    NotFound:
      description: Resource not found
    InternalServerError:
      description: An unexpected error occurred on the server
    Conflict:
      description: Conflict with the current state of the target resource
    BadRequest:
      description: Bad request body